- Add identity server to the project
- Fix formatting of PackageReference elements in IdentityServer.csproj for consistency
- Add initial migration files for AspNetUsers and update database context to use SQL Server
- Add authorization attributes and configure JWT authentication in controllers and Program.cs
- Add JWT authentication and update Program.cs for authorization
- Add Microsoft.AspNetCore.Authentication.JwtBearer package reference to project
- Add ConfigurationDbContext and related migrations for IdentityServer
- Add new client configuration and seed data for IdentityServer
- Update API scopes and allowed scopes for IdentityServer clients
- Add authorization policies for collection read and write operations
- Add authorization policies for advertisement and bid operations
- Add ValuationReadPolicy for authorization on valuation and brands endpoints
- Enhance error handling in WatchValuationHttpClient for brand and valuation retrieval
- Update comments in Config.cs to clarify client relationships and API scopes
- Fix exception handling status code for unavailable entities and implement access token retrieval in WatchValuationHttpClient
- Refactor IdentityServer configuration: update client ID, grant types, and allowed scopes; add CORS policy for SPA; change page title to WatchPlatform
- Add CORS policy to allow requests from specified origins
- Remove CORS policy configuration and add Email identity resource in Config.cs
- Implement ProfileService for user claims and roles; update Config.cs with roles identity resource; add CORS middleware in Program.cs; include Duende.IdentityModel package in project file.
- Reorder service registration for ProfileService and AspNetIdentity in HostingExtensions
- Update CORS settings and redirect URIs in Config.cs; disable inbound claims mapping in Program.cs
- Add authorization handler and requirement for claims and roles validation
- Update redirect URIs in Config.cs and adjust authorization policies in Program.cs for user role consistency
- Refactor AddBidAsync method to include bidder ID as a parameter and update Bid model accordingly
- Refactor advertisement service methods to include seller ID as a parameter for enhanced authorization checks
- Update authorization policies for bid deletion and add admin-specific policies
- Add TODO comment for potential user-specific advertisement retrieval
- Add comment to clarify admin-only bid deletion policy
- Refactor watch service and repository to include owner ID for enhanced authorization checks
- Enhance image upload and management by including owner ID for authorization checks
- Add additional redirect URIs and CORS origins for enhanced integration
- Update CORS policy to include new origin for enhanced security
- Add Azure Key Vault integration and update configuration settings
- Integrate Azure Key Vault for configuration management and update connection strings
- Update JWT Bearer authentication authority to production URL
- Update authentication aut...

+++ WatchPlatform/WatchCollection.Domain.Services/AdvertisementService.cs
@@ -12,14 +12,14 @@ namespace WatchCollection.Domain.Services;
 
 public class AdvertisementService(IAdvertisementRepository _repository, IWatchValuationHttpClient _watchValuationClient) : IAdvertisementService
 {
-    public async Task<AdvertisementResponseContract> CreateAdvertisement(AdvertisementRequestContract contract)
+    public async Task<AdvertisementResponseContract> CreateAdvertisement(string bidderIdString, AdvertisementRequestContract contract)
     {
         if(contract.Status is AdvertisementStatus.Sold || contract.Status is AdvertisementStatus.Expired)
             throw new InvalidAdvertisementStatusException("Cannot create an advertisement with status Sold or Expired.");
 
         var model = contract.AsModel();
         var advertisementId = Guid.NewGuid();
-        var sellerUserId = Guid.NewGuid(); // This should be retrieved from the authenticated user's context in a real application.
+        var sellerUserId = Guid.TryParse(bidderIdString, out var userIdGuid) ? userIdGuid : throw new Exception("Invalid User ID format.");
         model.AdvertisementId = advertisementId;
         model.SellerUserId = sellerUserId;
         model.ViewCount = 0;
@@ -29,16 +29,29 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
         return createdEntity.AsModel().AsContract();
     }
 
-    public async Task DeleteAdvertisement(Guid advertisementId)
+    public async Task DeleteAdvertisement(string sellerIdString, bool isAdmin, Guid advertisementId)
     {
+        var advertisement = await _repository.GetAdvertisementByIdAsync(advertisementId);
+        if (advertisement is null)
+            throw new AdvertisementNotFoundExceptions(advertisementId, "Advertisement not found");
+        if (advertisement.SellerUserId.ToString() != sellerIdString && !isAdmin){
+            // TODO: write custom exception for unauthorized access
+            throw new UnauthorizedAccessException("User is not authorized to delete this advertisement.");
+        }
         await _repository.DeleteAdvertisementAsync(advertisementId);
     }
 
-    public async Task<AdvertisementResponseContract?> GetAdvertisementById(Guid advertisementId)
+    public async Task<AdvertisementResponseContract?> GetAdvertisementById(string? userId, Guid advertisementId)
     {
         var entity =  await _repository.GetAdvertisementByIdAsync(advertisementId);
         if (entity is null)
             return null;
+        
+        if (userId is not null && entity.SellerUserId.ToString() != userId){
+            entity.ViewCount += 1;
+            await _repository.UpdateViewCountAsync(entity.AdvertisementId, entity.ViewCount);
+        }
+
         return entity.AsModel().AsContract();
     }
 
@@ -48,13 +61,19 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
         return entities.Select(e => e.AsModel().AsContract());
     }
 
-    public async Task<AdvertisementResponseContract> UpdateAdvertisement(Guid advertisementId, AdvertisementUpdateRequestContract contract)
+    public async Task<AdvertisementResponseContract> UpdateAdvertisement(Guid advertisementId, string sellerIdString, bool isAdmin, AdvertisementUpdateRequestContract contract)
     {
         if (contract.Status is AdvertisementStatus.Expired || contract.Status is AdvertisementStatus.Draft)
             throw new InvalidAdvertisementStatusException("Cannot update an advertisement to status Expired or Draft");
+
         var advertisement = await _repository.GetAdvertisementByIdAsync(advertisementId);
         if (advertisement is null)
             throw new AdvertisementNotFoundExceptions(advertisementId, "Advertisement not found");
+
+        if(advertisement.SellerUserId.ToString() != sellerIdString && !isAdmin)
+        {
+            throw new UnauthorizedAccessException("User is not authorized to update this advertisement.");
+        }
         
         var contractWithWatchId = new AdvertisementRequestContract
         {
@@ -66,8 +85,6 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
             AllowBids = contract.AllowBids
         };
         var model = contractWithWatchId.AsModel();
-        model.AdvertisementId = advertisementId;
-        model.SellerUserId = advertisement.SellerUserId; // This should be retrieved from the authenticated user's context in a real application.
 
         var entity = model.AsEntity();
